backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pymysql==1.1.1
cryptography==44.0.0
bcrypt==4.2.1
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.0
python-multipart==0.0.20
annotated-types==0.7.0
a2wsgi==1.10.7
